feat: read files contained in folder
what: read files contained in a folder as 3rd
argument.
Also add two optional parameters: limit and random
for a limit in the read images and a random read
of them as 4th and 5th arguments.
why: for an easy way to read image file instead
list.
how: this is the way to use this command now:
./resnet50_pt resnet50_DYB.xmodel [DYB_val] [random]