fix: finish tasks when queues are empty
what: when ordering inference from a large number
of images, there are missing images to process.
why: tasks are cascading out with items still in
the queue.
how: add a check to see if the queue is empty and
if the previous task is complete; if so, finish
the task.